Rename activity-viewer -> regulatory-activity-viewer
