+++ src/content/app/regulatory-activity-viewer/RegulatoryActivityViewer.tsx
