+++ src/content/app/regulatory-activity-viewer/components/activity-viewer-sidebar/ActivityViewerSidebar.module.css
