+++ src/content/app/regulatory-activity-viewer/components/region-overview/region-overview-image/RegionOverviewImage.module.css
