+++ src/content/app/regulatory-activity-viewer/components/region-overview/useRegionOverviewImageData.ts
@@ -17,7 +17,7 @@
 import type {
   OverviewRegion,
   RegulatoryFeature
-} from 'src/content/app/activity-viewer/types/regionOverview';
+} from 'src/content/app/regulatory-activity-viewer/types/regionOverview';
 
 /**
  * The purpose of this function is to group the region overview data into tracks.
@@ -37,16 +37,23 @@ type InputData = Pick<
 >;
 
 type Params = {
-  data: InputData;
+  data: InputData | null;
 };
 
-const prepareRegionOverviewTracks = (params: Params) => {
+// TODO: delete?
+
+const useRegionOverviewImageData = (params: Params) => {
   const { data } = params;
+
+  if (!data) {
+    return {
+      data: null
+    };
+  }
 
   const regulatoryFeatureTracks = buildRegulatoryFeatureTracks(data);
 
   return {
-    locations: data.locations,
     regulatoryFeatureTracks
   };
 };
@@ -77,6 +84,4 @@ const buildRegulatoryFeatureTracks = (input: NonNullable<Params['data']>) => {
   return regFeatureTracks;
 };
 
-export type PreparedTrackData = ReturnType<typeof prepareRegionOverviewTracks>;
-
-export default prepareRegionOverviewTracks;
+export default useRegionOverviewImageData;
